Refactored app running and migrations
- Created app factory to handle migrations and scrapers
- Moved migrations running from workflow file to dockerfile
- Created migrations.py to run migrations
- Changed app.py to call app factory
